refac buygd: move server/contract logic tosdk-v2
