+++ packages/good-design/src/apps/onramp/OnramperController.tsx
@@ -1,10 +1,8 @@
 import React, { useCallback, useEffect, useRef, useState } from "react";
 import { Onramper } from "./Onramper";
-import { useEthers, useEtherBalance, useTokenBalance, useCall } from "@usedapp/core";
+import { useEthers, useEtherBalance, useTokenBalance } from "@usedapp/core";
 import { WebViewMessageEvent } from "react-native-webview";
-import { useContractFunctionWithDefaultGasFees } from "@gooddollar/web3sdk-v2";
-import * as ethers from "ethers";
-import { Contract } from "ethers";
+import { swapHelper, useBuyGd } from "@gooddollar/web3sdk-v2";
 import { useModal } from "../../hooks/useModal";
 import { View, Text } from "native-base";
 import { WalletAndChainGuard } from "../../core";
@@ -12,48 +10,19 @@ import { useSignWalletModal } from "../../hooks/useSignWalletModal";
 
 export const OnramperController = () => {
   const cusd = "0x765de816845861e75a25fca122bb6898b8b1282a";
-  const { account, chainId, library } = useEthers();
+  const { account, library } = useEthers();
   const swapLock = useRef(false);
+
+  const { createAndSwap, swap, swapState, createState, gdHelperAddress } = useBuyGd();
 
   const { SignWalletModal } = useSignWalletModal();
 
-  // const buygdFactory = useGetContract("BuyGDFactory", false, "base");
-  const buygdFactory = new Contract("0x00e533B7d6255D05b7f15034B1c989c21F51b91C", [
-    "function createAndSwap(address owner,uint256 minAmount) external returns(address)",
-    "function predict(address owner) external view returns(address)"
-  ]);
-
-  const targetGDHelper = useCall(
-    account &&
-      buygdFactory && {
-        contract: buygdFactory,
-        method: "predict",
-        args: [account]
-      },
-
-    { refresh: "never", chainId }
-  );
-  const gdHelperAddress = targetGDHelper?.value?.[0];
-  const buygdInstance = new Contract(gdHelperAddress || ethers.constants.AddressZero, [
-    "function swap(uint256 minAmount,address gasRefund) external"
-  ]);
   const celoBalance = useEtherBalance(gdHelperAddress, { refresh: 1 });
   // const accountCeloBalance = useEtherBalance(account, { refresh: 1 }) || ethers.constants.Zero;
 
   const cusdBalance = useTokenBalance(cusd, gdHelperAddress, { refresh: 1 });
 
   const { showModal, Modal } = useModal();
-  const { send: swap, state: swapState } = useContractFunctionWithDefaultGasFees(buygdInstance, "swap", {
-    transactionName: "Exchange bought tokens to G$"
-  });
-
-  const { send: createAndSwap, state: createState } = useContractFunctionWithDefaultGasFees(
-    buygdFactory,
-    "createAndSwap",
-    {
-      transactionName: "Deploy secure swap helper"
-    }
-  );
 
   const selfSwap = false; //accountCeloBalance.gt(ethers.utils.parseUnits("5", 9).mul(400000)); //400000 gas for swap * 5gwei gas price
 
@@ -94,14 +63,12 @@ export const OnramperController = () => {
         console.log("swap tx res:", res);
         if (res?.status !== 1) throw Error("reverted");
       } else {
-        //or sbackends sends swap tx
-        const tx = fetch(`https://good-server.herokuapp.com/verify/swaphelper`, {
-          method: "POST",
-          headers: { "Content-type": "application/json" },
-          body: JSON.stringify({ account })
-        });
-        setStep(3);
-        await tx;
+        if (account) {
+          //or sbackends sends swap tx
+          const tx = swapHelper(account);
+          setStep(3);
+          await tx;
+        }
       }
       // when done set stepper at final step
       setStep(4);
