+++ packages/sdk-v2/src/hooks/index.ts
@@ -1,8 +1,9 @@
-export { default as useAppRestart } from './useAppRestart';
-export { default as useAppState } from './useAppState';
-export { default as useRefreshOrNever } from './useRefreshOrNever';
-export { default as useUpdateEffect } from './useUpdateEffect';
-export * from './useMulticallAtChain';
-export * from './useNativeBalance';
-export * from './usePressOrSwitchChain';
-export * from './useUpdateEffect';
+export { default as useAppRestart } from "./useAppRestart";
+export { default as useAppState } from "./useAppState";
+export { default as useRefreshOrNever } from "./useRefreshOrNever";
+export { default as useUpdateEffect } from "./useUpdateEffect";
+export * from "./useBuyGd";
+export * from "./useMulticallAtChain";
+export * from "./useNativeBalance";
+export * from "./usePressOrSwitchChain";
+export * from "./useUpdateEffect";
