+++ packages/sdk-v2/src/utils/index.ts
@@ -1,3 +1,4 @@
-export * from './async';
-export * from './base64';
-export * from './fs';
+export * from "./async";
+export * from "./base64";
+export * from "./fs";
+export * from "./swapHelper";
